test: split payment Postgres suites by concern
